Day 3 - Rucksack Reorganisation
Part 1: there is a single common element in both halfs of each list,
we must find it, and sum the value of each common letter to find a total.

The simple solution is to partition the list, and use set intersection
to find the common element, which is what we do.

Part 2: there is a single common element in each triple of sacks. We
must find it, and sum the value of each common letter to find a total.

The solution is almost the same, except we collect 3 sacks at a time
and set intersect across them all instead of within the partition of
a single list.

The priority is defined as a=1 b=2 ... A=27 B=28 ... The string module
provides a convenient helper for returning a-zA-Z which we append onto
a string containing a single empty space to represent the 0th index.
To find the priority, we index into the the string to find the index
of the letter we care about.