Introduce a batch util for processing iterables in batches
Batching through input is a common enough problem that I'm adding a
helper for it. Day 3 (2022) required batching 3 lines at a time, and I
initially used range with a step of N to achieve it. I knew something
in itertools would be able to help and I spent a few minutes trying
to find it.

But itertools doesn't have a native solution to the problem though it
does share a recipe, which I've now reproduced here.