Day 4 2022 completed
Part 1: process pairs of ranges, and count pairs that have a complete
overlap. We first check if LEFT fully contains RIGHT, then RIGHT fully
contains LEFT.

Part 2: process pairs of ranges, and acount pairs that have a partial
overlap. We convert both pairs to sets of ranges and check for an
intersection.