2022 Day 5 - Complete
Parse the input into a list of stacks and a tuple of instructions.

Part 1: process the instructions to move an item from the source to
the target stack one at a time.

Part 1: process the instructions to move N items from the source to
the target stack maintaining the order.