2022 Day 6
Find the index of the character after a given marker.

Part 1: the marker is found by finding the first 4 characters that are
all different. We use a sliding window to find all 4 character tuples,
then check the length of the set of those characters to see if they
are unique.

Part 2: the marker is found by finding the first 14 characters that are
all different. We use the same logic as part 1, but increase the length
check to 14.