2022 Day 7 - Complete
Parse a list of terminal output to determine the size of a tree of
directories so that we can delete enough files to make space for
an update.

Part 1: Find the recursive size of all directories with a size of at
most 100000 and sum them together.

Part 2: Find the recursive size of all directories and find the smallest
one that can be deleted to free up enough space required for the update.