2022 Day 9 - Complete
Instructions describe the movement of the head of a snake. When the
head moves, the rest of the body must be connected to it. When a gap
appears the body must move to fill the gap.

Part 1: the snake is of length two. The initial solution moved the head
and tail.

Part 2: the snake is of length ten. Refactoring the solution to iterate
over the entire body was required. Since the length was much longer,
bigger gaps could appear that required movement of more than 1 space,
so a while loop was introduced to progressively move each segment
closer.