2022 Day 8 Complete
Given a grid of tree heights determine the visibility of each tree
from a point of reference.

Part 1: count the number of visible trees from outside the grid. We
scan from the left, then the top, then from right and finally from the
bottom, counting only trees that are smaller than the max already found.

Scanning in this manner is O(n) as we visit each node just once.

Part 2: count the number of trees visible from each tree, stopping once
we hit a tree of same or greater height. This is wildly inefficient as
we rescan nodes over and over.

We could store the visibility in each direction for each node to avoid
some recomputation, but the solution is already found in a reasonable
amount of time so I'm moving on.