Day 11 2022 Complete
Monkeys are throwing items to each other. As each item is passed we
perform some calculations on it based on the attributes of the monkey.

Part 1: Perform the calculations particular to each monkey and reduce
the score by dividing by 3 and rounding down to an integer. Run 20
rounds and find the top 2 monkeys by the number of items they inspect.

Part 2: Perform the same calculations but do not reduce with floor
division. This time there are 10,000 rounds. As the number increases
the bottle neck becomes the modulo operation to check if the number
is divisable, and it will not complete in a reasonable time. We can
perform our own reduction by performing modulo the least common divisor
of each of the monkeys division operations. This keeps the item weight
within a reasonable bound and removes the previous bottleneck.