Tried an alternate for Part 2
The initial solution is slow. But, observing the final output, we can
see that the sand forms a 2-triangle (1,3,5,...). The maximum space it
can occupy is floor**2.

Observing the empty space we can see that a continuous horizontal line
of rocks also form a 2-triangle of empty space beneath them. If we can
count the rocks and scan each continuous horizontal wall and compute
their 2-triangles, we can subtract the rocks + empty space from the
maximum occupiable space.

Unfortunately there are edge cases that aren't hit for the test input
that mean this approach doesn't work naively.

The first is that non-continuous horizontal walls should sometimes act
as continuous if a previous empty triangle overlaps the gaps.

The second is when a full side of a triangle is boxed in it should
create more empty space past the first row.

These problems might be solveable but I've lost the will to finish.

The solution completes in under 1 second though, and dealing with the
edge cases would not significantly impact that.