Complete part 2 and optimise part 1
Part 1 uses the same original logic, but it massively reduces the
number of operations.

Part 1: For each sensor, scan the manhattan distance up and down and
find any that intersect the target row. At the intersection point,
compute the manhattan distance to that point, and then add the difference
both left and right. Eliminates all Point addition since we're only
checking along the x-axis.

Part 2: Uses a tool new to me, Z3, which is an equation/constraint
solver. We add all of the known constraints including min, max on the
grid, and exclude everything within the manhattan distance of each
scanner. It correctly finds the one point where the distress beacon
is.