Finishes Day 16
This solution uses almost the same approach as the original part 1,
but it compresses the graph such that we only care about going between
nodes that have opened valves.

We pre-compute the cost for each useful path and then DFS between each
of those nodes adding the cost as we go.

So, rather than iterating in all directions every minute, we can only
iterate between major nodes and skip all of the intermediate states.

Once we have all useful transitions we can find the max path or the max
two-distinct-paths and we're done.