2019 Day 7
This change refactors Input and Output so that multiple interpreters
can have their IO chained together.

When chained, a signal is raised when INPUT is requested or OUTPUT is
written to trigger the SENDER/RECEIVER to continue work.

We also introduce the concept of suspension. When writing output or
reading input, we suspend the interpreter so we prevent re-entrant
execution if a signal attempted to resume us recursively.